separate the sidebar pannel from auth slice
