+++ tabs/register.tsx
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from "react"
 import { Provider } from "react-redux"
 import { useSelector } from "react-redux"
 import { useDispatch } from "react-redux"
-import { checkCurrentUser, hideSidepanel, signUpWithEmail } from "../Slice/authSlice"
+import { checkCurrentUser, signUpWithEmail } from "../Slice/authSlice"
+import { hideSidepanel } from "../Slice/sidepanelSlice"
 import type { RootState, AppDispatch } from "../store"
 import { store } from "../store"
 import "../style.css"
