+++ tabs/reset-password.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react"
 import { Provider } from "react-redux"
 import { useDispatch } from "react-redux"
 import { sendResetEmail } from "../Slice/authSlice"
-import { hideSidepanel } from "../Slice/authSlice"
+import { hideSidepanel } from "../Slice/sidepanelSlice"
 import type { AppDispatch } from "../store"
 import { store } from "../store"
 import "../style.css"
@@ -14,6 +14,7 @@ function ResetPasswordPageContent() {
     // Hide sidepanel when on reset password page
     dispatch(hideSidepanel())
   }, [dispatch])
+
   const [email, setEmail] = useState("")
   const [error, setError] = useState("")
   const [message, setMessage] = useState("")
@@ -30,11 +31,14 @@ function ResetPasswordPageContent() {
     }
 
     setLoading(true)
-    const { error } = await sendResetEmail(email)
+    const resultAction = await dispatch(sendResetEmail(email) as any)
     setLoading(false)
 
-    if (error) setError(error.message)
-    else setMessage("Password reset link sent! Check your email.")
+    if (sendResetEmail.rejected.match(resultAction)) {
+      setError(typeof resultAction.payload === "string" ? resultAction.payload : "Reset failed")
+    } else {
+      setMessage("Password reset link sent! Check your email.")
+    }
   }
 
   const navigateToLogin = () => {
